AMLS-4166 | used moved status generator to generators package
